feat: :construction: finalizando service de registro e iniciando login

+++ app/backend/src/database/migrations/20221116183953-create-accounts.js
@@ -10,7 +10,7 @@ module.exports = {
         autoIncrement: true,
       },
       balance: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL,
         allowNull: false,
       }
     })
